Create a util package with common system, error and logging logic
